scribe/scribe-api/pyproject.toml
[project]
name = "scribe-api"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "httpx>=0.27",
    "python-jose[cryptography]>=3.3",
    "pydantic>=2.9",
    "pydantic-settings>=2.6",
    "asyncpg>=0.30",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "python-multipart>=0.0.12",
    "python-magic>=0.4",
    "pydub>=0.25",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.ruff]
line-length = 100
target-version = "py312"
